refactor: Move project-content related utils (LPS-119618)
