+++ src/import/import.js
@@ -7,8 +7,10 @@ const {
   FRAGMENT_IMPORT_STATUS,
   PAGE_TEMPLATE_IMPORT_STATUS,
 } = require('../utils/constants');
-const { default: getProjectContent } = require('../utils/get-project-content');
 const { log } = require('../utils/log');
+const {
+  default: getProjectContent,
+} = require('../utils/project-content/get-project-content');
 
 /**
  * Fragment types
