+++ src/utils/__tests__/write-project-content.spec.js
@@ -2,8 +2,10 @@ const fs = require('fs');
 const path = require('path');
 const tmp = require('tmp');
 
-const { default: getProjectContent } = require('../get-project-content');
-const writeProjectContent = require('../write-project-content');
+const {
+  default: getProjectContent,
+} = require('../project-content/get-project-content');
+const writeProjectContent = require('../project-content/write-project-content');
 
 describe('utils/write-project-content', () => {
   const projectCollections = JSON.parse(
